# refactor(gx2f): put parts into separate compile units