@paulgessinger 2024-12-05
One remaining issue with the formatting, aside from this looks good to merge.